Revert "Enable CNS swiftv2 path for SF/Standalone scenario to program secondary nics" (#2608)
* Revert "Enable CNS swiftv2 path for SF/Standalone scenario to program seconda…"

This reverts commit bc218edb4ef27593d0ccbabf01c7645dc16799d2.

* fix linter errors-1

* fix linter errors-2

* fix linter errors-3

* fix linter errors-4

* address comment, add todo
