+++ cns/NetworkContainerContract.go
@@ -465,7 +465,6 @@ type PodIpInfo struct {
 	PodIPConfig                     IPSubnet
 	NetworkContainerPrimaryIPConfig IPConfiguration
 	HostPrimaryIPInfo               HostIPInfo
-	HostSecondaryIPInfo             HostIPInfo
 	// NICType defines whether NIC is InfraNIC or DelegatedVMNIC or BackendNIC
 	NICType       NICType
 	InterfaceName string
@@ -475,15 +474,12 @@ type PodIpInfo struct {
 	SkipDefaultRoutes bool
 	// Routes to configure on interface
 	Routes []Route
-	// AddInterfacesDataToPodInfo is set to true for SF SwiftV2 scenario
-	AddInterfacesDataToPodInfo bool
 }
 
 type HostIPInfo struct {
-	Gateway     string
-	PrimaryIP   string
-	SecondaryIP string
-	Subnet      string
+	Gateway   string
+	PrimaryIP string
+	Subnet    string
 }
 
 type IPConfigRequest struct {
