+++ cns/api.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"net"
 	"time"
 
 	"github.com/Azure/azure-container-networking/cns/common"
@@ -364,36 +363,4 @@ type GetHomeAzResponse struct {
 type EndpointRequest struct {
 	HnsEndpointID string `json:"hnsEndpointID"`
 	HostVethName  string `json:"hostVethName"`
-}
-
-// GetEndpointResponse describes response from the The GetEndpoint API.
-type GetEndpointResponse struct {
-	Response     Response     `json:"response"`
-	EndpointInfo EndpointInfo `json:"endpointInfo"`
-}
-
-type EndpointInfo struct {
-	PodName       string
-	PodNamespace  string
-	IfnameToIPMap map[string]*IPInfo // key : interface name, value : IPInfo
-	HnsEndpointID string
-	HostVethName  string
-}
-type IPInfo struct {
-	IPv4 []net.IPNet
-	IPv6 []net.IPNet
-}
-
-type GetHTTPServiceDataResponse struct {
-	HTTPRestServiceData HTTPRestServiceData `json:"HTTPRestServiceData"`
-	Response            Response            `json:"response"`
-}
-
-// HTTPRestServiceData represents in-memory CNS data in the debug API paths.
-// PodInterfaceId is key and value is slice of Pod IP uuids in PodIPIDByPodInterfaceKey
-// secondaryipid(uuid) is key for PodIPConfigState
-type HTTPRestServiceData struct {
-	PodIPIDByPodInterfaceKey map[string][]string              `json:"podIPIDByPodInterfaceKey"`
-	PodIPConfigState         map[string]IPConfigurationStatus `json:"podIpConfigState"`
-	IPAMPoolMonitor          IpamPoolMonitorStateSnapshot     `json:"ipamPoolMonitor"`
 }
