+++ cns/client/client.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/Azure/azure-container-networking/cns"
+	"github.com/Azure/azure-container-networking/cns/restserver"
 	"github.com/Azure/azure-container-networking/cns/types"
 	"github.com/pkg/errors"
 )
@@ -562,7 +563,7 @@ func (c *Client) GetPodOrchestratorContext(ctx context.Context) (map[string][]st
 }
 
 // GetHTTPServiceData gets all public in-memory struct details for debugging purpose
-func (c *Client) GetHTTPServiceData(ctx context.Context) (*cns.GetHTTPServiceDataResponse, error) {
+func (c *Client) GetHTTPServiceData(ctx context.Context) (*restserver.GetHTTPServiceDataResponse, error) {
 	u := c.routes[cns.PathDebugRestData]
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String(), nil)
 	if err != nil {
@@ -582,7 +583,7 @@ func (c *Client) GetHTTPServiceData(ctx context.Context) (*cns.GetHTTPServiceDat
 	if res.StatusCode != http.StatusOK {
 		return nil, errors.Errorf("http response %d", res.StatusCode)
 	}
-	var resp cns.GetHTTPServiceDataResponse
+	var resp restserver.GetHTTPServiceDataResponse
 	err = json.NewDecoder(bytes.NewReader(b)).Decode(&resp)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to decode GetHTTPServiceDataResponse")
@@ -1023,7 +1024,7 @@ func (c *Client) GetHomeAz(ctx context.Context) (*cns.GetHomeAzResponse, error) 
 }
 
 // GetEndpoint calls the EndpointHandlerAPI in CNS to retrieve the state of a given EndpointID
-func (c *Client) GetEndpoint(ctx context.Context, endpointID string) (*cns.GetEndpointResponse, error) {
+func (c *Client) GetEndpoint(ctx context.Context, endpointID string) (*restserver.GetEndpointResponse, error) {
 	// build the request
 	u := c.routes[cns.EndpointAPI]
 	uString := u.String() + endpointID
@@ -1043,7 +1044,7 @@ func (c *Client) GetEndpoint(ctx context.Context, endpointID string) (*cns.GetEn
 		return nil, errors.Errorf("http response %d", res.StatusCode)
 	}
 
-	var response cns.GetEndpointResponse
+	var response restserver.GetEndpointResponse
 	err = json.NewDecoder(res.Body).Decode(&response)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to decode GetEndpointResponse")
