+++ cns/hnsclient/hnsclient_windows.go
@@ -27,9 +27,8 @@ const (
 	ExtHnsNetworkGwAddress = "192.168.255.1"
 
 	// HNS network types
-	hnsL2Bridge    = "l2bridge"
-	hnsL2Tunnel    = "l2tunnel"
-	hnsTransparent = "transparent"
+	hnsL2Bridge = "l2bridge"
+	hnsL2Tunnel = "l2tunnel"
 
 	// hcnSchemaVersionMajor indicates major version number for hcn schema
 	hcnSchemaVersionMajor = 2
@@ -138,7 +137,7 @@ func CreateDefaultExtNetwork(networkType string) error {
 		return nil
 	}
 
-	if networkType != hnsL2Bridge && networkType != hnsL2Tunnel && networkType != hnsTransparent {
+	if networkType != hnsL2Bridge && networkType != hnsL2Tunnel {
 		return fmt.Errorf("Invalid hns network type %s", networkType)
 	}
