+++ cns/restserver/restserver.go
@@ -103,6 +103,19 @@ type IPInfo struct {
 	IPv6 []net.IPNet
 }
 
+type GetHTTPServiceDataResponse struct {
+	HTTPRestServiceData HTTPRestServiceData `json:"HTTPRestServiceData"`
+	Response            Response            `json:"Response"`
+}
+
+// HTTPRestServiceData represents in-memory CNS data in the debug API paths.
+// TODO: add json tags for this struct as per linter suggestion, ignored for now as part of revert-PR
+type HTTPRestServiceData struct { //nolint:musttag // not tagging struct for revert-PR
+	PodIPIDByPodInterfaceKey map[string][]string                  // PodInterfaceId is key and value is slice of Pod IP uuids.
+	PodIPConfigState         map[string]cns.IPConfigurationStatus // secondaryipid(uuid) is key
+	IPAMPoolMonitor          cns.IpamPoolMonitorStateSnapshot
+}
+
 type Response struct {
 	ReturnCode types.ResponseCode
 	Message    string
@@ -136,7 +149,6 @@ type httpRestServiceState struct {
 	TimeStamp                        time.Time
 	joinedNetworks                   map[string]struct{}
 	primaryInterface                 *wireserver.InterfaceInfo
-	secondaryInterface               *wireserver.InterfaceInfo
 }
 
 type networkInfo struct {
