+++ cns/service/main.go
@@ -39,6 +39,7 @@ import (
 	nncctrl "github.com/Azure/azure-container-networking/cns/kubecontroller/nodenetworkconfig"
 	podctrl "github.com/Azure/azure-container-networking/cns/kubecontroller/pod"
 	"github.com/Azure/azure-container-networking/cns/logger"
+	"github.com/Azure/azure-container-networking/cns/middlewares"
 	"github.com/Azure/azure-container-networking/cns/multitenantcontroller"
 	"github.com/Azure/azure-container-networking/cns/multitenantcontroller/multitenantoperator"
 	"github.com/Azure/azure-container-networking/cns/restserver"
@@ -800,16 +801,6 @@ func main() {
 		if platform.HasMellanoxAdapter() {
 			go platform.MonitorAndSetMellanoxRegKeyPriorityVLANTag(rootCtx, cnsconfig.MellanoxMonitorIntervalSecs)
 		}
-		// if swiftv2 scenario is enabled, we need to initialize the Service Fabric (standalone) swiftv2 middleware to process IP configs requests
-		if cnsconfig.SWIFTV2Mode == configuration.SFSWIFTV2 {
-			cnsClient, err := cnsclient.New("", cnsReqTimeout) //nolint:govet // shadow ok as function returns in above errs
-			if err != nil {
-				logger.Errorf("Failed to init cnsclient, err:%v.\n", err)
-				return
-			}
-			swiftV2Middleware := &restserver.SFSWIFTv2Middleware{CnsClient: cnsClient}
-			httpRestService.AttachIPConfigsHandlerMiddleware(swiftV2Middleware)
-		}
 	}
 
 	// Initialze state in if CNS is running in CRD mode
@@ -1229,7 +1220,7 @@ func InitializeCRDState(ctx context.Context, httpRestService cns.HTTPService, cn
 
 	// check the Node labels for Swift V2
 	if _, ok := node.Labels[configuration.LabelNodeSwiftV2]; ok {
-		cnsconfig.SWIFTV2Mode = configuration.K8sSWIFTV2
+		cnsconfig.EnableSwiftV2 = true
 		cnsconfig.WatchPods = true
 		if nodeInfoErr := createOrUpdateNodeInfoCRD(ctx, kubeConfig, node); nodeInfoErr != nil {
 			return errors.Wrap(nodeInfoErr, "error creating or updating nodeinfo crd")
@@ -1413,13 +1404,22 @@ func InitializeCRDState(ctx context.Context, httpRestService cns.HTTPService, cn
 		}
 	}
 
-	if cnsconfig.SWIFTV2Mode == configuration.K8sSWIFTV2 {
+	if cnsconfig.EnableSwiftV2 {
 		if err := mtpncctrl.SetupWithManager(manager); err != nil {
 			return errors.Wrapf(err, "failed to setup mtpnc reconciler with manager")
 		}
 		// if SWIFT v2 is enabled on CNS, attach multitenant middleware to rest service
-		// here for AKS(K8s) swiftv2 middleware to process IP configs requests
-		swiftV2Middleware := &restserver.K8sSWIFTv2Middleware{Cli: manager.GetClient()}
+		// switch here for different type of swift v2 middleware (k8s or SF)
+		var swiftV2Middleware cns.IPConfigsHandlerMiddleware
+		switch cnsconfig.SWIFTV2Mode {
+		case configuration.K8sSWIFTV2:
+			swiftV2Middleware = &middlewares.K8sSWIFTv2Middleware{Cli: manager.GetClient()}
+		case configuration.SFSWIFTV2:
+		default:
+			// default to K8s middleware for now, in a later changes we where start to pass in
+			// SWIFT v2 mode in CNS config, this should throw an error if the mode is not set.
+			swiftV2Middleware = &middlewares.K8sSWIFTv2Middleware{Cli: manager.GetClient()}
+		}
 		httpRestService.AttachIPConfigsHandlerMiddleware(swiftV2Middleware)
 	}
